Added View menu and added Replace Indicators item and Period Type item under that menu.  Period type includes shortcuts so that the user can, e.g., switch from daily to weekly and back without having to use the mouse.  Removed the choice button on the left side of the window (and the left panel, since it is no longer needed) for selecting period type so that it is now selected only via the menu or by shortcut keys.
The MA_MenuBar class was created to allow moving a lot of the menu-related
code out of Chart.

Added Select Indicator item to the file menu to provide a pop-up list, similar
to the Select Market list, for choosing indicators.  Also added a shortcut for
this item so that the list can be popped up without having to use the mouse.

Moved the IndicatorListener class out from within Chart into its own file to
be a separate class.  It is now used by Chart and by IndicatorSelection, a new
class that is a listener that pops up the indicator selection list.