Some fairly massive changes, mostly to include handling of open interest and to change the scheme for updating and displaying the current indicator list so that the indicators displayed to the user are always appropriate (according to info. from the server) for the current tradable, the main case here being that data for a futures contract instrument will have open interest and indicators based on open interest, while stock data will not.  Removed the code in the constructor that calls data_builder.send_indicator_list_request to get the indicators for the first symbol.  The request_data method now handles calling data_builder.send_indicator_list_request - it calls it to update the indicator lists whenever the current market (symbol) changes to keep the indicator lists up to date.  The call to request_data in the constructor was moved to initialize_GUI_components, which works better with the new scheme - here is where the indicator list is first caused to be created for the first symbol on start up.
Changed `indicators' method to not recreate the indicator lists if the
old list received from the server is the same as the one just received.
Moved the workhorse code in `indicators' to a new routine.  Added changes
in the code to handle the Open interest field.  Added a MA_MenuBar attribute,
ma_menu_bar and call its (new) update_indicators method when `indicators'
recreates the indicator lists.

Added a new_indicators attribute that is set by request_data when it sends
a request for a new indicator list; request_data then calls `indicators',
which was changed to recreate the indicator lists if new_indicators is true
or if _indicators is null (has not been set yet) and the old list and new
list are not equal (as mentioned above).  The new attribute,
old_indicators_from_server, is used to hold the old list of indicators
received from the server.

There may have been a couple of other minor changes, which I have
forgotten about.