Added a new open_interest query - is there an open interest field in the current data?  Added a call in send_market_data_request to the new routine setup_parser_fieldspecs, which resets the parser's field specs to correspond to the current data attributes - currently simply whether or not there is an open interest field.  Added _last_open_interest attribute that serves the same purpose for OI as the last volume attribute does for volume.
In new_main_drawer, for some reason java 1.3 was not liking the use of
the c.Candle_graph, etc., constants in a switch/case statement, so changed
the scheme to use "if/else"s instead.

Added initialize_fieldspecs, which initializes the field specs that are given
to the Parser when it is created.  It handles the new Open_interest case.

In initialize, changes to call the new initialize_fieldspecs and to handle the
new open interest case.

...