Fixed bug in send_response in the new code to check for open interest and add the OI flag to the preface if it's there: if t.has_open_interest then     pref := concatenation(<<clone(pref), Open_interest_flag>>) end
t can be Void and cause a segmentation violation.  Added a "t /= Void"
condition in the if to prevent that.

In send_response, moved check for server_error first in the error-checking
section - after determining that tradable is Void.  This fixes a bug - if
the guard that calls a tradables method is called before server_error,
which relied on the current state of tradables, the error state of
tradables will be reset and thus no error will be detected.  Checking
server_error first ensures that if a server error occurred (such as not
being able to open the file for the current tradable), it will be detected
with the if statement and reported in the statement within the if block.