Changes to the interface to allow the client to have more control over which tradables will be processed, instead of the old scheme where the event generator has the tradable dispenser and operates on its current item.  Now the EG takes a tradable to process, via the new routines set_tradable_from_dispenser and set_tradable_from_pair. The tradable to be processed by the `execute' routine is "extracted" when these routines are called, either from the current item of the dispenser (set_tradable_from_dispenser) or the appropriate member of the "tradable pair" (an intraday and non-intraday data set for the same tradable), according to the EG's period type, passed in by set_tradable_from_pair. This scheme allows a client to do event processing on just one tradable (which can be used, for example, to obtain trading signals for one stock in response to a request from a GUI client).
Added a new ancestor, MARKET_PROCESSOR, and implemented its queries -
functions, operators, and parameters.  This allows a client to determine
things about the event generator, such as whether it uses (directly or
indirectly) any operators that access an open-interest field (which is not
valid for stocks).  These queries may also allow easier access to information
related to operators and functions and may allow more complete editing
capabilities for event generators.