Added new edit_minus_n procedure to handle editing new n_adjustment attribute of the minus-n command, as well as the old attributes.
Added edit_index_extractor to handle editing the new INDEX_EXTRACTOR
command.

In record_lowest_offset, changed to use new `external_offset' feature
of the LINEAR_OFFSET_COMMAND, instead of `offset'.  The new feature
allows the offset to not be adjusted to externally (by providing an
external offset value of 0), which is needed with some linear offset
commands.