In command_types, added a new list element for INDEXED commands.  (See the new INDEX_EXTRACTOR.)  And added the INDEX_EXTRACTOR as a real-result command.
Added a constant for the new INDEXED type; added a Minus_n constant for
MINUS_N_COMMAND, which now gets its own editing classification (updated in
initialization_map and initialize_command).  And added an "Index" constant
for the new editing classification for the new INDEX_EXTRACTOR command -
and made corresponding changes for this classification in
initialization_map and initialize_command.