Added an external_offset feature to allow a client's view of the offset (which may sometimes be 0 when the offset is not 0) to differ from the actual offset value.  Changed the export status of `offset' to private - it's now only used internally.
The external offset change was made to accommodate the need to create
a weighted moving average function that uses a MINUS_N_COMMAND whose
offset is n + 1, but whose external offset is 0.  (The set_effective_offset
procedure used by APPLICATION_FUNCTION_EDITOR needs to, in the WMA case,
ignore the minus-n command's offset because the function is constructed
such that the target list has the cursor advanced such that the minus-n
command will not access any position of target whose index is < 1; and,
of course, using an effective offset here messes up the resulting
calculation.)  It's likely that this flexibility will be needed for
other functions in the future.