Finalization of the further abstraction of the function used for SMA and standard deviation (but kept the old one, for now - sma_based_calculation) into a new n_based_accumulation function, which is more configurable - instead of addition (for the "accumulation" operator) and subtraction (for the "drop-off" operator), a binary operator is passed in to perform these operations, so that other calculations, such as product, can be performed instead of a SUM.
Had forgotten to plug in 'accum_op' in the 'binary_double_operation' call -
Fixed that.  Did some cleanup - removal of test code, etc. and added some
explanatory, and reminder, comments.  More cleanup is needed.

I tested this version with standard deviation - against the old std. dev.
implementation - the result was correct (the same as the result with the
old version) and the timing (with an input data set size of 1000 daily
records [/tmp/ibm.txt]) was essentially the same, or even slightly faster,
though probably not significantly so, for the new version.

Because of it works correctly and there is no performance hit over the
old version, I will probably decide to make this new version (i.e,.,
using the n_based_accumulation function) the "official" one.