The change to use launch instead of system causes the shell command to fail because the input file being used for the email is removed ('msg_file.delete', at the end of 'notify_by_email') before the shell command is actually executed, causing a "No such file ..." error message.  Changed back to use 'system' to solve this problem.
Note: In the future it may be appropriate to replace the mailing
functionality here with a better, more sophisticated "event-delivery"
system of some kind, which would make it easier to implement a working
solution on Windows and to allow more options for "event-message" delivery.

Also, improved the ugly use of " <" in the mail command by placing it into
a string constant - not a complete solution, but should make it easier to
adapt the construct for other OSs if needed.