Finished the missing functionality - parsing the command line, using regular expressions to convert user responses into object names and convert back when replaying, etc.  And added a good bit of debugging code to find a bug that only occurs when assertions are off - the system dies (e.g., when processing -i indicators) at the same place every time.  When running in the debugger, it suddenly uses a large amount of memory and then hangs. It was reproduceable but not diagnosable.
I found the source of the problem by first determining that the bug does
not occur when postcondition checking is on.  Then I did a binary search,
changing the ace file to recursively turn postconditions off for the
current suspected clusters until I found the cluster where the problem was
occurring.  Then I eliminated classes in that cluster until I narrowed it
down to the KL_STRING_ROUTINES class, in gobo's kernel/spec/ise.  I then
used a similar procedure to narrow the problem down to the routine:
append_substring_to_string.  I then experimented until I found a "fix"
for the problem where the bug would not occur when postconditions are off.
The diff for the fix is:

799a800
> 			wk_string: STRING
805,808c806,812
< 				from i := s until i > e loop
< 					a_string.append_character (other.item (i))
< 					i := i + 1
< 				end
---
> 				wk_string := clone (a_string) + other.substring (s, e)
> 				a_string.clear_all
> 				a_string.append (wk_string)
> --				from i := s until i > e loop
> --					a_string.append_character (other.item (i))
> --					i := i + 1
> --				end

(I think this is a generated class, so I saved a copy of the fixed version
in /opt/eiffel/gobo/library/kernel/fixes/spec/ise.)