Configuration-file changes:
Allow platform-specific (windows vs. unix) settings - currently just
the directory separator: '/' is always used in the mctrc file as the
separator, but in windows it is changed during processing to a '\'.
Needed to add a concept of escape character (currently '~'); other
related changes.

Also added the the environment_variable specification to allow setting an
environment variable to occur as part of processing the config. file.

Fixed at least one bug uncovered during the work above.