Added a new project for an "extended" MAS server that will allow the release of a commercial version with features not available in the "open-source" (or free or whatever it will be) version.  Created an extensions cluster under server1 and subclusters within it for classes that provide "extended" behavior (including a separate root class that sets up, polymorphically, some of the "extended" objects).
The first "extended" feature is a "report-back" feature: The MCT asks the
server to report, on startup, whether it started successfully and, if not,
what went wrong.  The server sends a message, via a socket connection, back
to the MCT with this information and the MCT displays it if there is an
error or starts up a control window for the server if there was no error.
(The MCT does not have an "extensions" cluster.  It may be appropriate to
add a setting in the mctrc file specifying whether the report-back feature
is available in the server.  If it's not, the MCT would not request a
"report-back".)