Removed reminder note: this has already been done.
Changes to allow starting a session when MCT starts up and starting up the
charting process automatically when a new mas session is started.  This is
likely to be a highly requested feature.  I may leave this as the default.
Note that functionality needs to be added to allow the user (from the MCT
GUI) to turn this on or off - Currently they have to turn it off by
removing the start_... settings in the mctrc file or setting them to false.

Note that I've discovered what can be considered a bug.  (I thought it
might have been caused by these changes, but I tested with an old version
of MCT and the same problem occurs.)  The bug is that if you start up
several mas sessions, at some point the server hangs.  It's usually
(perhaps always - I tried testing several times to see if this would happen
in the 2nd session and it never did.) the 3rd session or later where this
occurs.  I tested starting the (lastest version of the) mas server from the
command-line starting up 12 processes in a for loop:

for pn in 1234 1235 1236 1237 1238 1239 1231 1232 1233 12345 12346 12347
do
./mas_no_assert_v1.6.6p -f , -b /opt/mas/lib/data/tradables/*.txt $pn &
done

and the problem did not occur, so it only appears when the MCT is used.
I suspect there's a problem caused by the MCT/MAS-server communication that
occurs when the server is started up, perhaps related to timing.