Some fairly extensive changes to improve the formatting, such as bringing the list of links down to the left of the bullet list and expanding the main text to the left (shrinking the left margin), separating into more paragraphs, etc.  A lot of these changes resulted from suggestions from Tania.
I also put pretty much the entire page in a table and created <td></td>
sections, which fixed the stubborn problem of putting the SF logos at the
bottom, across the whole page.

The next step is to reformat it, probably using tables, to keep the same
structure, in order to allow it to work in Netscape 4.