I accidentally fed cvs the wrong file for the checkin comments last time. This is what should have been submitted.  (Also made a couple minor changes to allow this checkin to happen.)
Some formatting/layout changes: Added a "header" section; made the side
navigation bar into a visible bar by making the bg a different color and
extending it to 100% of the window.  A small content/edit.

The header in its current form does not look good or function well, so I
will remove it; but I'm checking it in here in case remnants of it are
needed (for ideas or whatever) in the future.