For a "server error", changed the type of error reported to the client from an "Error" to a "Warning".  This fixes the following problem:
When the -w option is used and the client is set up to request a symbol
that is in the 'symbols' file but that doesn't exist (i.e., the request to
yahoo.com fails), reporting an Error caused the client to treat the error
as fatal and exit immediately.  Reporting a warning allows the client to
display the error message in a window and continue.

NOTE: This change might introduce a new problem: If the server error
actually is fatal, the warning response in the client - to keep going -
may be innappropriate.