Improved the scheme for the 'command-line framework':  COMMAND_LINE.make now calls prepare_for_argument_processing, for preprocessing, before calling process_arguments (main_setup_procedures) and then calls finish_argument_processing, for postprocessing. The descendant is expected to implement prepare_for_argument_processing, main_setup_procedures, and finish_argument_processing.
process_arguments is implemented to call each procedure in its argument
(list) continuously until the query last_argument_found is false.
This allows the procedure to be implemented to set last_argument_found
to true to force all duplicate command-line options processed by that
procedure to be processed and removed.  The previous implementation
would regard a duplicate option (e.g., mas -p 12345 -p) as an invalid
option and fail.