Fixed the following MCT bug:
When using the "connect to a running mas session" option with the MCT to
connect to a remote machine, 'start cl client' does not work - mct is
starting macl with the wrong host name (localhost instead of the given
remote host).

The problem was that the <hostname> tag was getting replaced by the
'hostname' setting on start up, which works for the 'start a MAS session'
option, but not for connecting to a 'running mas session' on a different
machine from that specified by the 'hostname' setting.  The fix was to
eliminate the start-up processing of the hostname tag and, just like the
port number, let it be processed when the command to connect to a new
session is run.  In this case the <hostname> tag is replaced by the host
name of the window for the session, which will be the one the user typed
into the 'locate mas session' window.

I also had to add 'Hostname_specifier' to the 'dynamic_tokens' array so
that 'check_for_unreplaced_tokens' will not mark the <hostname> tag as an
undefined token.

Tested successfully for both the standard 'start a MAS session' option
and the 'connect to a running mas session' option.

Debugging print statements need to be removed.