Some refactoring and other changes to synchronize the support 'package' in the old and new GUIs to allow both projects to share the same 'support' files.  Among other things, this involved taking the axis-related code out of IndicatorGroup and putting it into a new class: MonoAxisIndicatorGroup. (The new GUI gets a BiAxisIndicatorGroup - with two axes.)
The new GUI version of MA_Configuration has diverged enough that it may not
be worht it to refactor that class.  Instead, I'll probably split it into
two versions - one for the old gui and one for the new gui.  It will need
to be moved into its own directory/package - application_support might be a
good name.

After this check in, I need to make the above MA_Configuration change,
check in the changed files in the new GUI (on Pedro's machine), let Pedro
know that we need to remove the cvs archive of the support files on his
machine so there's only one version (, etc.).