Moved MA_Configuration to a new package - application_support, to break the dependency of the support packag on it.  Introduced a parent of MA_Configuration to allow this break of dependency: MA_ConfigurationInterface.
Too bad Java forces the developer to put import statements into the code -
It causes a mess when a class is moved, such as the MA_Configuration move.
Added the necessary import statements.

Made small changes in a couple files (such as CompressedConnection) so that
they match the version on Pedro's machine, in preparation for putting them
under the control of one cvs archive.

Changed ParameterBasedConfigurationModifier to use the new
MA_ConfigurationInterface instead of MA_Configuration to break the
above-mentioned dependency.

Etc.