Fix for task ID 28: "mas crashes when time period is too small" (description: Problem reported by Orest: When using the charting GUI against intraday data with a small time period (such as 5-minute), swtitching to a tradable whose smallest period is larger (such as 15-minute) appears to cause the server to crash. As reported by Orest: "MAS will still crash if you try to graph in 5 and 10 minute mode--(because I am only putting out 15 minute data).")
I located the cause of the problem: The GUI received an 'invalid period type'
error from the server when it, e.g., asks for a 5-minute period type for a
tradable whose smallest type is 15-minutes.

Implemented a fix: Added a new message ID in the server: Invalid_period_type

Server now sends that ID instead of Error when it gets a request for
an invalid period type.

Changed the GUI to recognize the Invalid_period_type message ID and requery
the server for the valid types for the current tradable; then it requests
the smallest (in the case of intraday data) type available.

I also changed the code to 'reset' the period_type menu to the currently
available types.

One possible complaint people might have about this solution is that once
the smaller list of period types is set up (the invalid types are
discarded), the GUI continues to use this list, even for tradables that
have smaller types available.  (However, all the types can be obtained by
opening a new gui window.)  A possible enhancement would be to fix this
"problem".

Note that this problem appeared in Orest's data because of its funky
structure - the first data set - bond data - gets a period type of 5
minutes because of the first two records of the data.  The data is (using
my term) not "normalized".  It would probably be better from the client's
point of view to normalize the data to 15 minutes.

However, I think this is a valid problem to fix because other customers
could run into a case where the data is normalized, but some tradables have
different period types available.