Refactoring: Extracted ChartFacilities and ChartInterface classes to allow some facilities to be shared by the old (free) and new (commercial) GUIs.
The immediate reason for doing this is to allow the changes for the
time-period bug (description below) to be shared by the old and new GUIs.
Here's the description of the bug:

Problem reported by Orest: When using the charting GUI
against intraday data with a small time period (such as
5-minute), swtitching to a tradable whose smallest period
is larger (such as 15-minute) causes the GUI to crash. As
reported by Orest: "MAS will still crash if you try to
graph in 5 and 10 minute mode--(because I am only putting
out 15 minute data)."

NOTE1: A few new routines lack an implementation in the new GUI.  These are
marked with !!! comments.

NOTE2: For the new GUI: The new send_data_request routine needs to be
implemented for the period-types fix to work.  It appears that according
to the new GUI structure, this functionality no longer belongs in this
class.  This may not be a problem - e.g., delegation may be used to produce
a relatively clean design.  If it is a problem we will either need to
design a new solution or abandong the shared ChartFacilities scheme.

And Here are relevant cvs entries (for Chart.java) that give more detail on
what changes were made - may be useful for figuring out what is going
on with the new GUI:

revision 2.80
date: 2004/07/28 18:35:48;  author: jtc;  state: Exp;  lines: +44 -8
Extended the previous change wrt period types to do the following if an
invalid period type situation is encountered (handle_invalid_period_type
is called):

Treat each tradable separately wrt period types, so that all available
period types for a tradable are accessible in the GUI.  E.g.:

   tradable X has smallest period type of 5 minute
   tradable Y has smallest period type of 15 minute
   tradable Z has smallest period type of 1 minute

Once the "separate handling" state is triggered (see
the 'individual_period_type_sets' variable in the code), deal with X, Y,
and Z separately such that all available (from the server) period types are
also available in the GUI, but period type that is invalid for a particular
tradable is ever available.
----------------------------
revision 2.79
date: 2004/07/24 04:25:11;  author: jtc;  state: Exp;  lines: +8 -4
Changed handle_invalid_period_type to not try again to request data
if the period types list returned by the server is empty.
----------------------------
revision 2.78
date: 2004/07/24 01:14:58;  author: jtc;  state: Exp;  lines: +19 -0
Fix for task ID 28: "mas crashes when time period is too small"
(description: Problem reported by Orest: When using the charting GUI
against intraday data with a small time period (such as
5-minute), swtitching to a tradable whose smallest period
is larger (such as 15-minute) appears to cause the server
to crash. As reported by Orest: "MAS will still crash if
you try to graph in 5 and 10 minute mode--(because I am
only putting out 15 minute data).")

I located the cause of the problem: The GUI received an 'invalid period type'
error from the server when it, e.g., asks for a 5-minute period type for a
tradable whose smallest type is 15-minutes.

Implemented a fix: Added a new message ID in the server: Invalid_period_type

Server now sends that ID instead of Error when it gets a request for
an invalid period type.

Changed the GUI to recognize the Invalid_period_type message ID and requery
the server for the valid types for the current tradable; then it requests
the smallest (in the case of intraday data) type available.

I also changed the code to 'reset' the period_type menu to the currently
available types.

One possible complaint people might have about this solution is that once
the smaller list of period types is set up (the invalid types are
discarded), the GUI continues to use this list, even for tradables that
have smaller types available.  (However, all the types can be obtained by
opening a new gui window.)  A possible enhancement would be to fix this
"problem".

Note that this problem appeared in Orest's data because of its funky
structure - the first data set - bond data - gets a period type of 5
minutes because of the first two records of the data.  The data is (using
my term) not "normalized".  It would probably be better from the client's
point of view to normalize the data to 15 minutes.

However, I think this is a valid problem to fix because other customers
could run into a case where the data is normalized, but some tradables have
different period types available.