Fix for the following bug:
When the data values are all the same, the data is invisible on the chart
(because it is drawn at the top or bottom).  Fix: If the ymax/ymin values
are the same add a constant to ymax and subtract the same constant from
ymin so that the "height" of the graph screen is greater than 0.  With the
fix, the data is drawn as a horizontal line in the middle of the graph.