Some improvements to tradable_list: Turned the cache into a queue so that when the cache is full the oldest tradable is dropped off of the cache (instead of the old method of just clearing the cache when it gets full).
Changed the meaning of 'caching on' to be that the cache-queue capacity > 0.
This means the 'caching on' is no longer a settable state that has to be
checked for.  The first 'if' statement in 'item' now only needs to check if
the cursor (index) changed.

Tested with database, files, and web data sources (as well as with large
intraday files) to verify that it behaved as expected.