Searched for who calls 'clear_caches' and, unsurprisingly, found that the class I'm interested in is mas_login_request_cmd.  Since I don't want the cache to be cleared when the auto-update feature is available, I made the call to 'clear_caches' conditional upon a new constraint: not auto_update_on - i.e., clear the cache iff auto_update_on is False.
I then added a global_configuration class and, for the open-source
version, basic_global_configuration class with an auto_update_on setting
of false, and, for the non-OS version, an extended_global_configuration
calss with an auto_update_on setting of true, using a factory to select
which class to use.  I then modified main_gui_interface to use this
global-config.auto_udpate feature when creating the mas_login_request_cmd
object.