Changes related to the auto-data-update task (34), mostly to the tradable_list hierarchy.  Changed the algorithm in item, in tradable_list, to call new routines - target_tradable_out_of_date and append_new_data, to update the current tradable if the data is out of date.
Provided implementations of the above 2 routines and supporting routines
in file_tradable_list - update the data with the current file only if the
file's mod. date and size has changed.

Made some changes to allow the data-scanning/input framework to load the
data from the current file (or input medium) cursor position
instead of calling 'start' to start at the beginning.  This allows the
system to only read the new data and append it to the tradable's data
array.  These changes required some relatively small changes (like
removing the call to 'start') in the eiffel library classes INPUT_FILE
and DATA_SCANNER.

Tested the file-based data loading, though it should be tested more.

There is more work to do - refactoring, moving the enhancements to the
"extended" classes, cleaning up, etc.  This check in allows saving the
changes so that they don't get lost - before these cleanups are done.

Some other related changes and improvements.