Date-related improvements:
Use the new date_now feature.

Use a future date (100 years from "now") as an "eternal now" (always > the
latest tuple date).

When finding the index of the last tuple not later than the specified end
date (in 'last_index' of market_tuple_printer), if the end date is in
the future and the last tuple's date is not, simply provide the index of
the last tuple rather than the more expensive call to
l.index_at_date.

Added a debug type in the ace files.