The change in the data scanner to not call input.start caused a bug in code that expected input.start to be called.  (It showed up in market analysis.) Instead of changing the code to call input.start externally to the data scanner, I changed the interface to data scanner to include a start_input and let the state default to true by using an implementation attribute do_not_start_input_before_scanning that gets initialized to false so that descendants' create routines do not need to do anything.
Changed tradable factory to include the same input-start state and set the
state on its data scanner accordingly.  And in 'append_new_data' of
extended_file_based_tradable_list the input-start state is turned off so
that input.start is not called.  Everywhere else, the default of calling
input.start is left on.