Moved the attributes min_max_epsilon min_border_boundary from DataSet to BasicDrawer.  The fix to draw data sets for which the value is all the same in the middle of the screen instead of at the top (which was in DataSet) is now in 'draw_data' of BasicDrawer.  Tested successfully.
I also removed duplicate attributes (attributes in the parent, Drawer, that
were redeclared in BasicDrawer), and associated set_... routines.

Note: I believe that the Java spec. says that these attributes (redeclared
in the child) do not redefine the inherited attributes (as with Eiffel),
but create new attributes with the same name.  This is, of course, a
very poor decision on the part of the Java language designers, leading
to bugs such as the following in BasicDrawer.java, which I "fixed"
earlier by putting set_... routines back in:

revision 2.17
date: 2004/08/25 22:33:41;  author: jtc;  state: Exp;  lines: +12 -0
The previous changes I made to this file - removing routine redefinitions
that were simply a duplicate of the implementation in the parent -
apparently cause a bug: The upper indicators are drawn in a different scale
than the main data, resulting in the indicator line not being lined up
right with the data/candles.  I don't know why the change causes this bug,
and don't feel it's worth the time to investigate at this point, so I
restored the file to its state before the previous changes.  ...


Of course, I now know that the reason removing these 'duplicate' routines
(set_... routines) caused a problem was that the associated attributes were
also duplicated.  And due to Java's poor design, compiling this file gives
no error or warning, simply allowing the attributes to be duplicated.
Removing the set_... routines resulted in the parent set_... routine being
called, which set the parent's 'x' attribute rather than the child's.
Why the fuck does Java allow this?

Here is the diff from revisions 2.17 to 2.16 of BasicDrawer.java that
shows what routines were removed and then put back in to fix the bug:

Index: BasicDrawer.java
===================================================================
RCS file: /home/cvs/finance/src/clients/graph/BasicDrawer.java,v
retrieving revision 2.17
retrieving revision 2.16
diff -r2.17 -r2.16
51,62d50
< 	public void set_xaxis(Axis a) {
< 		xaxis = a;
< 	}
<
< 	public void set_yaxis(Axis a) {
< 		yaxis = a;
< 	}
<
< 	public void set_clipping(boolean b) {
< 		clipping = b;
< 	}
<

Now that the duplicate attributes xaxis, yaxis, and clipping have been
removed and the associated set_... routines have been "re-removed", the
problem should no longer exist.