Deleted:
//import java.applet.*;

I may need to split the file into two versions (or find another solution by
refactoring) at some point if it turns out that the open-source gui, which
does not rely on swing, runs into problems because of its inheritance from
JApplet.  (The most likely problem is that windows users are stuck with a
java implementation that does not support swing, due to Microsoft's idiocy,
and that some of these users are not willing to take an action such as
installing Sun's JRE.)  If this happens, the open-source version will
inherit from Applet and the commercial version will inherit from Japplet.