The last checkin for this file was an accident - a slip using the cim (cvs) command caused the contents of mas-extended.ace to be submitted as the log entry.
This entry is the one that should have been used for the previous checkin.

The previous bug fix (a few hours ago) was not adequate.  Instead,
I implemented a better fix in tradable_list.  However, I'm leaving the
"old fixes" (pertaining to the cache) in this file because I think they
add useful features and a useful, appropriate invariant condition.

The change for this checkin is just a comment describing the checkin.