A bit of refactoring: Added a new class: UNARY_MATH_OPERATOR and changed log*.e and square_root.e to inherit from it - allows removal of some duplicated code.
Added a SINE command that inherits from UNARY_MATH_OPERATOR.  Followed the
procedure in
/home/development/jtc/finance/src/server1/doc/procedure_for_adding_a_command
to add the SINE operator to the system.

Compiled and tested - ran into a bug, which I've fixed.
(The bug was introduced during the
addition of the extended_file_tradable_list and related classes and
was not discovered before because I did not test the new version of the
system with respect to the path for creating a new indicator, which was
used for this test.)

After the bug fix, I created a market function that provides the sine of
the volume, and verified on a test data set that the output was correct.