Some fairly extensive changes to try to make time period types immutable and, essentially, n-pletons - so that only one of each period type resides in the system, maintained by the new period type facilities class.  (This class is now the only class allowed to make a TIME_PERIOD_TYPE.)
This change also involves adding the concept of a non-standard period type
(e.g, 3-minute, or 6-minute), based on the fact that some data sets (such
as Orest's) will not be normal (1-min, 2-min, 5-min, or etc.), but rather
either odd, such as 3- or 4-min, or uneven ("un-normalized") such that
using an abnormal period type for the base period type may be the best
solution to produce an acceptable result.

A new 'allow non-std period types' option in the command line specifies
that if the data makes it necessary to use such a non standard type, that
the odd type is not presented to the user unless this option is true.
Instead, it is used to build standard types - e.g., a 3-minute base type
for a tradable yields a 5-minute type as the lowest that the user will
see for that tradable.

The above appears to be the best scheme for handling strange data.
However, the current implementation takes 'allow non-std period types' to
mean that the odd types should be discarded and, instead, the next larger
standard type should be used - e.g., if the data merits a 4-minute
duration, use 5-minute instead.  However, this solution doesn't work as
well as the above - e.g., with (essentially) 4-minute data a 5-minute type
is used as the base type, but when the "5-minute" data is printed, it is
the original "4-minute" data, since the base period type data is not
constructed - it is used as is - the raw data unchanged.

With the former solution (that will be implemented in the next checkin),
in the above example the 4-minute type will be used and 5-minute data will
be constructed from the 4-minute data and will thus be "more correct".
And if 'allow non-std period types' is true, the 4-minute raw data will be
available for viewing; otherwise it will not be, but it will be used
internally to construct the 5-minute, ... data.

I've added some --!!! comments noting where to make changes to implement
the "proposed" solution.