More work on the immutable TIME_PERIOD_TYPEs enhancement issue:
Implemented the changes re. non-standard period types that I documented in
the previous check-in.  Made the planned changes to TRADABLE_LIST_HANDLER -
Reimplemented period_type_names_for to check the command-line option
allow_non_standard_period_types and, if it is not set, to remove the
"non-standard" period types from the selection list that is presented
to the user.  Added supporting enhancements to PERIOD_TYPE_FACILITIES -
e.g., the standard_period_types query (with a side-track of using the
new standard_period_type_indexes function to recode sections that used
to use the one_minute, two_minute, etc. indexes explicitly - in order
to shorten and clean up the code).

INTRADAY_TUPLE_DATA_SCANNER - removed the use of
allow_non_standard_period_types.

Ran some tests on the sample "4-minute" data file I've been using the past
couple days (/tmp/dtst1.15-minute) and confirmed that the behavior is as
expected: With 'allow_non_standard_period_types' set, the "4-minute" period
type is available for selection, both with macl and with the charting GUI.
With 'allow_non_standard_period_types' set to false, the "4-minute" period
is not available for selection.  In both cases, the 4-minute time period
type is used internally as the base period type and used to construct the
5-minute, ... period data.