Brought build.xml and mas.jsp up to date with the scheme currently being used on the test platform: io:/usr/java/jarkarta*/webapps/mas.
Remove use of mas_client.jar - use maslib.jar instead.

maslib.jar is now placed in the war file in WEB-INF/lib.  The new file:

scripts/deploy

"deploys" the jar file by copying it into the webapp directory and links
maslib.jar into ./classes so that the applet can use it - load from it.
This script needs to be changed to use the paths that will be used on
Orest's machine.